try adding auth.config
