+++ routiner/src/middleware.ts
@@ -1,2 +1,5 @@
 export { auth as middleware } from "@/auth";
-export const config = { matcher: ["/board", "/calendar"] }
+export const config = { 
+    matcher: ["/board", "/calendar"],
+    runtime: 'nodejs',
+};
